requirements.txt
flask==3.1.0
torch
torchvision
transformers
Pillow
supabase
python-dotenv
